requirements.txt
streamlit==1.40.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.1
pyyaml==6.0.2
python-dotenv==1.0.1
tenacity==9.0.0
rich==13.9.4
typer==0.13.1
sqlalchemy==2.0.36
supabase==2.9.1
aiofiles==24.1.0
mmh3==5.0.1
pytest==8.3.4
pytest-asyncio==0.25.0
